fix tappable areas in my collection form
