+++ src/lib/Scenes/MyCollection/Screens/ArtworkFormModal/Components/__tests__/ArrowDetails-tests.tsx
@@ -1,0 +1,18 @@
+import { renderWithWrappers } from "lib/tests/renderWithWrappers"
+import { ArrowRightIcon } from "palette"
+import React from "react"
+import { View } from "react-native"
+import { ArrowDetails } from "../ArrowDetails"
+
+describe("ArrowDetails", () => {
+  it("renders correct components", () => {
+    const Noop: React.FC = () => <View />
+    const wrapper = renderWithWrappers(
+      <ArrowDetails>
+        <Noop />
+      </ArrowDetails>
+    )
+    expect(wrapper.root.findByType(Noop)).toBeDefined()
+    expect(wrapper.root.findByType(ArrowRightIcon)).toBeDefined()
+  })
+})
